+++ src/ModelConstructors.jl
@@ -55,6 +55,10 @@ module ModelConstructors
         AbstractVectorParameter, VectorParameter, VectorParameterVector, ScaledVectorParameter,
         UnscaledVectorParameter, Untransformed,
 
+        # regimes.jl
+        set_regime_val!, regime_val, set_regime_prior!, regime_prior, set_regime_fixed!, regime_fixed,
+        set_regime_valuebounds!, regime_valuebounds, toggle_regime!,
+
         # statistics.jl
         prior, posterior, posterior!,